Simple for validation methodology. More testing with sessions.
I have designed a simple form validation method that may or may not stay the same
with the usage of json. You have the form struct and a Validate interface. This makes
form validtion rather easy.

I am messing with sessions more to understand them fully. Next steps are to generate
random strings for the name of a new session (like when logging in) and store that
random string while also associating it with a user. From there you can set the values
about authenticated status and other sensitive things in the database reducing
security risks. This may mean we can just use the secure cookie package.