Worked on formatting data from database to json appropriate for the client.
We need to have a discussion now about how we are presenting the data.
Currently presentation and storage are in conflict. The interface is too
limiting in how much you can see and searching to show all things tagged
a certain way is not obvious. The current data model is more complicated
than the presentation allows.